+++ Pet-Clinic/Pet-Clinic-Data/src/main/java/com/abhi/PetClinic/Service/OwnerService.java
@@ -4,7 +4,7 @@ import java.util.Set;
 
 import com.abhi.PetClinic.Model.Owner;
 
-public interface OwnerService extends CrudService<Owner, Integer> {
+public interface OwnerService extends CrudService<Owner, Long> {
 
 	Owner findByLastName(String lastName);
