+++ Pet-Clinic/Pet-Clinic-Data/src/main/java/com/abhi/PetClinic/Service/PetService.java
@@ -4,6 +4,6 @@ import java.util.Set;
 
 import com.abhi.PetClinic.Model.Pet;
 
-public interface PetService extends CrudService<Pet, Integer> {
+public interface PetService extends CrudService<Pet,Long> {
 
 }
